requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
unstructured
pysqlite3-binary
